repairedRVs
checking the proper functionality of activitites
