+++ app/src/main/AndroidManifest.xml
@@ -1,19 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
+
     <uses-permission android:name="android.permission.CALL_PHONE" />
     <uses-permission android:name="android.permission.SEND_SMS" />
 
-
     <queries>
-    <intent>
-        <action android:name="android.intent.action.CALL_BUTTON"/>
-    </intent>
-    <intent>
-        <action android:name="android.intent.action.SENDTO"/>
-    </intent>
+        <intent>
+            <action android:name="android.intent.action.CALL_BUTTON" />
+        </intent>
+        <intent>
+            <action android:name="android.intent.action.SENDTO" />
+        </intent>
+    </queries>
 
-</queries>
     <application
         android:allowBackup="true"
         android:dataExtractionRules="@xml/data_extraction_rules"
@@ -23,18 +23,24 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.TutorFinder"
-        tools:targetApi="31">
+        tools:targetApi="31" >
         <activity
-            android:name=".student.TutorProfileActivity"
-            android:exported="false">
+            android:name=".student.TutorFilterActivity"
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
-
+        <activity
+            android:name=".student.TutorProfileActivity"
+            android:exported="false" >
+            <meta-data
+                android:name="android.app.lib_name"
+                android:value="" />
+        </activity>
         <activity
             android:name=".student.MyTutorProfileActivity"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
@@ -42,53 +48,45 @@
         <activity
             android:name=".tutor.EditTutorProfileActivity"
             android:exported="false" >
-        <meta-data
-            android:name="android.support.PARENT_ACTIVITY"
-            android:value=".tutor.TutorScreenActivity" />
-        </activity>
-        <activity
-            android:name=".student.FilterTutorActivity"
-            android:exported="false">
             <meta-data
-                android:name="android.app.lib_name"
-                android:value="" />
+                android:name="android.support.PARENT_ACTIVITY"
+                android:value=".tutor.TutorScreenActivity" />
         </activity>
         <activity
             android:name=".student.EditStudentProfile"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
             android:name=".student.StudentDashboard"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
             android:name=".SignUpActivity"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
             android:name=".tutor.TutorScreenActivity"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.app.lib_name"
                 android:value="" />
         </activity>
         <activity
             android:name=".tutor.RequestedStudentProfileActivity"
-            android:exported="false">
+            android:exported="false" >
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
                 android:value=".tutor.StudentRequestActivity" />
         </activity>
-
         <activity
             android:name=".tutor.MyStudentProfileActivity"
             android:exported="false" />
@@ -105,29 +103,28 @@
         <activity
             android:name=".MainActivity"
             android:exported="true"
-            android:theme="@style/NoActionBar">
+            android:theme="@style/NoActionBar" >
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-
             <intent-filter>
                 <action android:name="android.intent.action.SENDTO" />
+
                 <data android:scheme="mailto" />
+
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
-
-
             <intent-filter>
                 <action android:name="android.intent.action.SEND" />
+
                 <data android:scheme="sms" />
                 <data android:scheme="smsto" />
                 <data android:mimeType="text/plain" />
+
                 <category android:name="android.intent.category.DEFAULT" />
             </intent-filter>
-
-
         </activity>
 
         <meta-data
